+++ SWP391_CareSkin_BE/DTOS/Responses/Promotion/PromotionDTO.cs
@@ -6,7 +6,7 @@ namespace SWP391_CareSkin_BE.DTOS.Responses
     {
         public int PromotionId { get; set; }
         public string Name { get; set; }
-        public decimal DiscountPercent { get; set; }
+        public double DiscountPercent { get; set; }
         public DateTime StartDate { get; set; }
         public DateTime EndDate { get; set; }
         public bool IsActive { get; set; }
